Removed reply from action, clean up code

+++ src/components/Action.tsx
@@ -2,8 +2,9 @@ import { ActionLabel } from "./Model/ActionModel";
 import "../styles/Action.scss";
 import { collection, deleteDoc, getDocs } from "firebase/firestore";
 import { db } from "../Firebase";
-import { useSelector } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import { RootState } from "../redux/store";
+import { updateReplyMessage } from "../redux/slices/replyMessageSlice";
 
 function Action({
   actions,
@@ -16,16 +17,22 @@ function Action({
     receiverId: string;
   };
 }) {
-
   // ************ Variables ************
   let height = actions.length * 20;
-
 
   // *********** Redux ***************
 
   const currentConvId = useSelector(
     (state: RootState) => state.currentConvId.id
   );
+
+  const currentReply = useSelector(
+    (state: RootState) => state.reply.replyMessage
+  );
+
+  const dispatch = useDispatch();
+
+  // ************* Functions
 
   // To delete message entry
   const deleteMessageEntry = () => {
@@ -44,10 +51,26 @@ function Action({
     );
   };
 
+  const replyMessageEntry = () => {
+    dispatch(
+      updateReplyMessage({
+        message: infos.content,
+        senderName: "",
+      })
+    );
+    console.log("After update: ", currentReply);
+  };
+
+  // **************** Rendering ************************
+
   return (
     <div
       className="dropdown"
-      style={{ height: height.toString() + "px", transition: "opacity 750ms", opacity: "0" }}
+      style={{
+        height: height.toString() + "px",
+        transition: "opacity 750ms",
+        opacity: "0",
+      }}
       onMouseLeave={(e) => {
         e.currentTarget.style.opacity = "0";
       }}
@@ -56,10 +79,11 @@ function Action({
         return (
           <button
             id={action.label}
-           onClick={() => {
-            if (action.label === "Delete")
-              deleteMessageEntry();
-          }}>
+            onClick={() => {
+              if (action.label === "Delete") deleteMessageEntry();
+              else if (action.label === "Reply") replyMessageEntry();
+            }}
+          >
             <action.icon />
             {action.label}
           </button>
